# kontrak/openapi.yaml
openapi: 3.1.0
info:
  title: Satu Muatan API
  description: Sistem konsolidasi muatan & bukti mutu rantai pasok hortikultura.
  version: 0.1.0
paths:
  /api/auth/masuk:
    post:
      tags:
      - auth
      summary: Masuk
      description: Login nomor HP + PIN 6 digit (§9.1).
      operationId: masuk_api_auth_masuk_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MasukRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/masuk-demo:
    post:
      tags:
      - auth
      summary: Masuk Demo
      description: 'Masuk cepat (demo) — hanya aktif saat DEMO_MODE (§9.1, K6: 6 akun).'
      operationId: masuk_demo_api_auth_masuk_demo_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MasukDemoRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/auth/saya:
    get:
      tags:
      - auth
      summary: Saya
      description: Profil pengguna yang sedang login.
      operationId: saya_api_auth_saya_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PenggunaOut'
      security:
      - HTTPBearer: []
  /api/komoditas:
    get:
      tags:
      - master
      summary: Daftar Komoditas
      operationId: daftar_komoditas_api_komoditas_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KomoditasOut'
                type: array
                title: Response Daftar Komoditas Api Komoditas Get
      security:
      - HTTPBearer: []
  /api/penerima:
    get:
      tags:
      - master
      summary: Daftar Penerima
      operationId: daftar_penerima_api_penerima_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PenerimaOut'
                type: array
                title: Response Daftar Penerima Api Penerima Get
      security:
      - HTTPBearer: []
  /api/koperasi/saya:
    get:
      tags:
      - master
      summary: Koperasi Saya
      description: Koperasi milik pengguna login (gudang = titik awal rute).
      operationId: koperasi_saya_api_koperasi_saya_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KoperasiOut'
      security:
      - HTTPBearer: []
  /api/konfigurasi:
    get:
      tags:
      - panel-asumsi
      summary: Daftar Konfigurasi
      operationId: daftar_konfigurasi_api_konfigurasi_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/KonfigurasiOut'
                type: array
                title: Response Daftar Konfigurasi Api Konfigurasi Get
      security:
      - HTTPBearer: []
  /api/konfigurasi/{kunci}:
    patch:
      tags:
      - panel-asumsi
      summary: Ubah Konfigurasi
      operationId: ubah_konfigurasi_api_konfigurasi__kunci__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: kunci
        in: path
        required: true
        schema:
          type: string
          title: Kunci
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KonfigurasiPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KonfigurasiOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tier-kendaraan:
    get:
      tags:
      - panel-asumsi
      summary: Daftar Tier
      operationId: daftar_tier_api_tier_kendaraan_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TierKendaraanOut'
                type: array
                title: Response Daftar Tier Api Tier Kendaraan Get
      security:
      - HTTPBearer: []
  /api/tier-kendaraan/{tier_id}:
    patch:
      tags:
      - panel-asumsi
      summary: Ubah Tier
      operationId: ubah_tier_api_tier_kendaraan__tier_id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: tier_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Tier Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TierKendaraanPatch'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TierKendaraanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/permintaan:
    get:
      tags:
      - permintaan
      summary: Daftar Permintaan
      description: 'Ter-scope per peran (K6): PENERIMA → miliknya; KOPERASI → semua
        yang terbuka.'
      operationId: daftar_permintaan_api_permintaan_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PermintaanOut'
                type: array
                title: Response Daftar Permintaan Api Permintaan Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - permintaan
      summary: Buat Permintaan
      operationId: buat_permintaan_api_permintaan_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermintaanCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermintaanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /api/slot/{slot_id}/lot:
    get:
      tags:
      - lot
      summary: Daftar Lot Slot
      description: Daftar lot sebuah slot — layar Muat (§9.5).
      operationId: daftar_lot_slot_api_slot__slot_id__lot_get
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LotOut'
                title: Response Daftar Lot Slot Api Slot  Slot Id  Lot Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/lot/{lot_id}/muat:
    patch:
      tags:
      - lot
      summary: Muat Lot
      description: Timbang + foto + checkbox 'Ada cacat terlihat' (input kunci atribusi
        §6).
      operationId: muat_lot_api_lot__lot_id__muat_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: lot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lot Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MuatPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/selesai-muat:
    post:
      tags:
      - lot
      summary: Selesai Muat
      description: Selesai muat → slot JALAN, waktu berangkat tercatat (§9.5).
      operationId: selesai_muat_api_slot__slot_id__selesai_muat_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LotOut'
                title: Response Selesai Muat Api Slot  Slot Id  Selesai Muat Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/lot/masuk:
    get:
      tags:
      - lot
      summary: Lot Masuk
      description: '''Pilih dari daftar'' (§9.7, K6): lot menuju penerima login yang
        belum

        diserahterimakan — jalur demo teraman.'
      operationId: lot_masuk_api_lot_masuk_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/BuktiLotOut'
                type: array
                title: Response Lot Masuk Api Lot Masuk Get
      security:
      - HTTPBearer: []
  /api/lot/qr/{kode_qr}:
    get:
      tags:
      - lot
      summary: Bukti Lot
      description: 'Bukti lot dari scan QR: foto muat, berat, waktu, transit berjalan
        vs ambang.'
      operationId: bukti_lot_api_lot_qr__kode_qr__get
      security:
      - HTTPBearer: []
      parameters:
      - name: kode_qr
        in: path
        required: true
        schema:
          type: string
          title: Kode Qr
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuktiLotOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/lot/{lot_id}/serah-terima:
    post:
      tags:
      - lot
      summary: Serah Terima
      description: Terima / Terima dengan potongan / Tolak → atribusi + PENJELASAN
        (§6, §9.7).
      operationId: serah_terima_api_lot__lot_id__serah_terima_post
      security:
      - HTTPBearer: []
      parameters:
      - name: lot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Lot Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SerahTerimaCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SerahTerimaOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/pengiriman:
    get:
      tags:
      - lacak
      summary: Pengiriman Slot
      description: Timeline + estimasi tiba (dari ambang transit) + jejak posisi.
      operationId: pengiriman_slot_api_slot__slot_id__pengiriman_get
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PengirimanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/pengiriman/{pengiriman_id}/majukan:
    post:
      tags:
      - lacak
      summary: Majukan Pengiriman
      description: Majukan state simulasi MockVendor satu langkah (K5) — deterministik,
        untuk demo.
      operationId: majukan_pengiriman_api_pengiriman__pengiriman_id__majukan_post
      security:
      - HTTPBearer: []
      parameters:
      - name: pengiriman_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Pengiriman Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PengirimanOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/dampak/ringkasan:
    get:
      tags:
      - dampak
      summary: Dampak Ringkasan
      description: 4 kartu, masing-masing dengan rumus + status_sumber. Tanpa data
        = null → '—'.
      operationId: dampak_ringkasan_api_dampak_ringkasan_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DampakRingkasanOut'
      security:
      - HTTPBearer: []
  /api/dampak/bulanan:
    get:
      tags:
      - dampak
      summary: Dampak Bulanan
      description: Grafik batang per bulan (Recharts) — termasuk jumlah_kiriman (K6).
      operationId: dampak_bulanan_api_dampak_bulanan_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/DampakBulananOut'
                type: array
                title: Response Dampak Bulanan Api Dampak Bulanan Get
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    BuktiLotOut:
      properties:
        lot:
          $ref: '#/components/schemas/LotOut'
        durasi_transit_berjalan_menit:
          anyOf:
          - type: integer
          - type: 'null'
          title: Durasi Transit Berjalan Menit
        ambang_transit_menit:
          type: integer
          title: Ambang Transit Menit
        serah_terima:
          anyOf:
          - $ref: '#/components/schemas/SerahTerimaOut'
          - type: 'null'
      type: object
      required:
      - lot
      - ambang_transit_menit
      title: BuktiLotOut
      description: Respons GET /api/lot/qr/{kode_qr} — layar Serah Terima (§9.7).
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KomoditasOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nama:
          type: string
          title: Nama
        satuan:
          type: string
          title: Satuan
        harga_acuan_per_kg:
          type: integer
          title: Harga Acuan Per Kg
        umur_simpan_jam:
          type: integer
          title: Umur Simpan Jam
        laju_susut_per_jam:
          type: number
          title: Laju Susut Per Jam
        status_sumber:
          $ref: '#/components/schemas/StatusSumber'
        catatan_sumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Sumber
      type: object
      required:
      - id
      - nama
      - satuan
      - harga_acuan_per_kg
      - umur_simpan_jam
      - laju_susut_per_jam
      - status_sumber
      title: KomoditasOut
    KonfigurasiOut:
      properties:
        kunci:
          type: string
          title: Kunci
        nilai:
          type: string
          title: Nilai
        tipe:
          $ref: '#/components/schemas/TipeKonfigurasi'
        label:
          type: string
          title: Label
        satuan:
          anyOf:
          - type: string
          - type: 'null'
          title: Satuan
        status_sumber:
          $ref: '#/components/schemas/StatusSumber'
        catatan_sumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Sumber
      type: object
      required:
      - kunci
      - nilai
      - tipe
      - label
      - status_sumber
      title: KonfigurasiOut
    KonfigurasiPatch:
      properties:
        nilai:
          type: string
          title: Nilai
      type: object
      required:
      - nilai
      title: KonfigurasiPatch
    KoperasiOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nama:
          type: string
          title: Nama
        kode:
          anyOf:
          - type: string
          - type: 'null'
          title: Kode
        desa:
          anyOf:
          - type: string
          - type: 'null'
          title: Desa
        kecamatan:
          anyOf:
          - type: string
          - type: 'null'
          title: Kecamatan
        kabupaten:
          anyOf:
          - type: string
          - type: 'null'
          title: Kabupaten
        alamat_gudang:
          type: string
          title: Alamat Gudang
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
      type: object
      required:
      - id
      - nama
      - alamat_gudang
      - lat
      - lng
      title: KoperasiOut
...
    LotOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kode_qr:
          type: string
          title: Kode Qr
        partisipasi_id:
          type: string
          format: uuid
          title: Partisipasi Id
        nama_petani:
          type: string
          title: Nama Petani
        nama_komoditas:
          type: string
          title: Nama Komoditas
        volume_kg:
          type: integer
          title: Volume Kg
        penerima_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Penerima Id
        nama_penerima:
          anyOf:
          - type: string
          - type: 'null'
          title: Nama Penerima
        berat_aktual_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Berat Aktual Kg
        foto_muat:
          anyOf:
          - type: string
          - type: 'null'
          title: Foto Muat
        waktu_muat:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Waktu Muat
        catatan_muat:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Muat
        cacat_terlihat:
          type: boolean
          title: Cacat Terlihat
      type: object
      required:
      - id
      - kode_qr
      - partisipasi_id
      - nama_petani
      - nama_komoditas
      - volume_kg
      - cacat_terlihat
      title: LotOut
...
    MasukDemoRequest:
      properties:
        akun:
          $ref: '#/components/schemas/AkunDemo'
      type: object
      required:
      - akun
      title: MasukDemoRequest
    MasukRequest:
      properties:
        no_hp:
          type: string
          title: No Hp
          examples:
          - 081234567001
        pin:
          type: string
          maxLength: 6
          minLength: 6
          title: Pin
          examples:
          - '123456'
      type: object
      required:
      - no_hp
      - pin
      title: MasukRequest
    MuatPatchRequest:
      properties:
        berat_aktual_kg:
          type: integer
          exclusiveMinimum: 0.0
          title: Berat Aktual Kg
        foto_muat_base64:
          anyOf:
          - type: string
          - type: 'null'
          title: Foto Muat Base64
        cacat_terlihat:
          type: boolean
          title: Cacat Terlihat
          default: false
        catatan_muat:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Muat
      type: object
      required:
      - berat_aktual_kg
      title: MuatPatchRequest
...
    PenerimaOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nama:
          type: string
          title: Nama
        tipe:
          $ref: '#/components/schemas/TipePenerima'
        alamat:
          type: string
          title: Alamat
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
      type: object
      required:
      - id
      - nama
      - tipe
      - alamat
      - lat
      - lng
      title: PenerimaOut
    PenggunaOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nama:
          type: string
          title: Nama
        no_hp:
          type: string
          title: No Hp
        peran:
          $ref: '#/components/schemas/PeranPengguna'
        koperasi_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Koperasi Id
        penerima_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Penerima Id
      type: object
      required:
      - id
      - nama
      - no_hp
      - peran
      title: PenggunaOut
...
    PeranPengguna:
      type: string
      enum:
      - PETANI
      - KOPERASI
      - PENERIMA
      title: PeranPengguna
    PermintaanCreate:
      properties:
        komoditas_id:
          type: string
          format: uuid
          title: Komoditas Id
        volume_kg:
          type: integer
          exclusiveMinimum: 0.0
          title: Volume Kg
        tanggal_dibutuhkan:
          type: string
          format: date
          title: Tanggal Dibutuhkan
      type: object
      required:
      - komoditas_id
      - volume_kg
      - tanggal_dibutuhkan
      title: PermintaanCreate
    PermintaanOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        penerima_id:
          type: string
          format: uuid
          title: Penerima Id
        nama_penerima:
          type: string
          title: Nama Penerima
        komoditas_id:
          type: string
          format: uuid
          title: Komoditas Id
        nama_komoditas:
          type: string
          title: Nama Komoditas
        volume_kg:
          type: integer
          title: Volume Kg
        volume_terpenuhi_kg:
          type: integer
          title: Volume Terpenuhi Kg
        tanggal_dibutuhkan:
          type: string
          format: date
          title: Tanggal Dibutuhkan
        status:
          $ref: '#/components/schemas/StatusPermintaan'
        slot_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Slot Id
        dibuat_pada:
          type: string
          format: date-time
          title: Dibuat Pada
      type: object
      required:
      - id
      - penerima_id
      - nama_penerima
      - komoditas_id
      - nama_komoditas
      - volume_kg
      - volume_terpenuhi_kg
      - tanggal_dibutuhkan
      - status
      - dibuat_pada
      title: PermintaanOut
...
    SerahTerimaCreate:
      properties:
        keputusan:
          $ref: '#/components/schemas/KeputusanSerahTerima'
        persen_potongan:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Persen Potongan
          default: 0
        alasan:
          anyOf:
          - type: string
          - type: 'null'
          title: Alasan
        foto_bongkar_base64:
          anyOf:
          - type: string
          - type: 'null'
          title: Foto Bongkar Base64
      type: object
      required:
      - keputusan
      title: SerahTerimaCreate
    SerahTerimaOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        lot_id:
          type: string
          format: uuid
          title: Lot Id
        penerima_id:
          type: string
          format: uuid
          title: Penerima Id
        waktu_bongkar:
          type: string
          format: date-time
          title: Waktu Bongkar
        keputusan:
          $ref: '#/components/schemas/KeputusanSerahTerima'
        persen_potongan:
          type: integer
          title: Persen Potongan
        alasan:
          anyOf:
          - type: string
          - type: 'null'
          title: Alasan
        durasi_transit_menit:
          type: integer
          title: Durasi Transit Menit
        ambang_transit_menit:
          type: integer
          title: Ambang Transit Menit
        atribusi:
          $ref: '#/components/schemas/Atribusi'
        penjelasan:
          type: string
          title: Penjelasan
      type: object
      required:
      - id
      - lot_id
      - penerima_id
      - waktu_bongkar
      - keputusan
      - persen_potongan
      - durasi_transit_menit
      - ambang_transit_menit
      - atribusi
      - penjelasan
      title: SerahTerimaOut
...
    TierKendaraanOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kode:
          type: string
          title: Kode
        nama:
          type: string
          title: Nama
        kapasitas_kg:
          type: integer
          title: Kapasitas Kg
        tarif_dasar:
          type: integer
          title: Tarif Dasar
        tarif_per_km:
          type: integer
          title: Tarif Per Km
        urutan:
          type: integer
          title: Urutan
        aktif:
          type: boolean
          title: Aktif
        status_sumber:
          $ref: '#/components/schemas/StatusSumber'
        catatan_sumber:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Sumber
      type: object
      required:
      - id
      - kode
      - nama
      - kapasitas_kg
      - tarif_dasar
      - tarif_per_km
      - urutan
      - aktif
      - status_sumber
      title: TierKendaraanOut
    TierKendaraanPatch:
      properties:
        kapasitas_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Kapasitas Kg
        tarif_dasar:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tarif Dasar
        tarif_per_km:
          anyOf:
          - type: integer
          - type: 'null'
          title: Tarif Per Km
        aktif:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Aktif
      type: object
      title: TierKendaraanPatch
...
    TokenResponse:
      properties:
        token:
          type: string
          title: Token
        pengguna:
          $ref: '#/components/schemas/PenggunaOut'
      type: object
      required:
      - token
      - pengguna
      title: TokenResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer